requirements.txt
GitPython==3.1.40
openai>=1.0.0
anthropic>=0.7.0
vijil>=0.1.0
